Add xml char iteration to BXMLItem
