+++ bvm/ballerina-core/src/main/java/org/ballerinalang/model/values/BXMLCodePointIterator.java
@@ -1,0 +1,54 @@
+/*
+ * Copyright (c) 2019, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * you may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.ballerinalang.model.values;
+
+/**
+ * {@link BXMLCodePointIterator} private iteration provider for BXML char sequence.
+ *
+ * @since 0.995.0
+ */
+public class BXMLCodePointIterator implements BIterator {
+    private String charSequence;
+    private int offset;
+
+
+    public BXMLCodePointIterator(String charSequence) {
+        this.charSequence = charSequence;
+        this.offset = 0;
+    }
+
+    static BXMLCodePointIterator from(String seq) {
+        return new BXMLCodePointIterator(seq);
+    }
+
+    @Override
+    public BValue getNext() {
+        int codePoint = charSequence.codePointAt(offset);
+        offset += Character.charCount(codePoint);
+
+        // Max 2 chars per code point.
+        StringBuilder sb = new StringBuilder(2);
+        sb.appendCodePoint(codePoint);
+        return new BString(sb.toString());
+    }
+
+    @Override
+    public boolean hasNext() {
+        return offset < charSequence.length();
+    }
+}
