+++ bvm/ballerina-core/src/main/java/org/ballerinalang/model/values/BXMLItem.java
@@ -841,6 +841,7 @@ public final class BXMLItem extends BXML<OMNode> {
 
         BXMLItem value;
         int cursor = 0;
+        BXMLCodePointIterator codePointIterator;
 
         BXMLItemIterator(BXMLItem bxmlItem) {
             value = bxmlItem;
@@ -848,7 +849,13 @@ public final class BXMLItem extends BXML<OMNode> {
 
         @Override
         public BValue getNext() {
-            if (hasNext()) {
+            if (value.getNodeType() == XMLNodeType.TEXT) {
+                if (codePointIterator == null) {
+                    codePointIterator = new BXMLCodePointIterator(value.stringValue());
+                }
+                cursor++;
+                return codePointIterator.getNext();
+            } else if (hasNext()) {
                 cursor++;
                 return value;
             }
@@ -857,6 +864,12 @@ public final class BXMLItem extends BXML<OMNode> {
 
         @Override
         public boolean hasNext() {
+            if (value.getNodeType() == XMLNodeType.TEXT) {
+                if (codePointIterator == null) {
+                    codePointIterator = new BXMLCodePointIterator(value.stringValue());
+                }
+                return codePointIterator.hasNext();
+            }
             return cursor == 0;
         }
